Rename base package to pscdiscrepanciesapi
