+++ src/test/java/uk/gov/ch/pscdiscrepanciesapi/PscDiscrepancyApiApplicationTests.java
@@ -1,6 +1,5 @@
-package uk.gov.ch.pscdiscrepancyapi;
+package uk.gov.ch.pscdiscrepanciesapi;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
